Fix circular import during type checking
