+++ src/zenml/utils/downloaded_repository_context.py
@@ -20,8 +20,8 @@ from zenml.code_repositories import LocalRepositoryContext
 class _DownloadedRepositoryContext(LocalRepositoryContext):
     """Class that represents the downloaded files of a code repository.
 
-    See `source_utils.set_custom_local_repository(...)` for a more in-depth
-    explanation.
+    See `code_repository_utils.set_custom_local_repository(...)` for a more
+    in-depth explanation.
     """
 
     def __init__(
